+++ src/in-memory-backend.service.ts
@@ -1,5 +1,6 @@
 import { Inject, OpaqueToken, Optional } from '@angular/core';
-import { BaseResponseOptions, Headers, Request, RequestMethod, Response, ResponseOptions } from '@angular/http';
+import { BaseResponseOptions, Connection, Headers, ReadyState, Request, RequestMethod,
+         Response, ResponseOptions, URLSearchParams } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import { Observer }   from 'rxjs/Observer';
 import 'rxjs/add/operator/delay';
@@ -80,6 +81,7 @@ export interface ReqInfo {
   collectionName: string;
   headers: Headers;
   id: any;
+  query: URLSearchParams;
   resourceUrl: string;
 }
 
@@ -93,21 +95,21 @@ export const isSuccess = (status: number): boolean => (status >= 200 && status <
  *
  * ### Usage
  *
- * Create InMemoryDataService class the implements IInMemoryDataService.
+ * Create InMemoryDataService class the implements InMemoryDataService.
  * Register both this service and the seed data as in:
  * ```
  * // other imports
- * import { HTTP_PROVIDERS, XHRBackend } from 'angular2/http';
- * import { InMemoryBackendConfig, InMemoryBackendService, SEED_DATA } from '../in-memory-backend/in-memory-backend.service';
+ * import { HTTPPROVIDERS, XHRBackend } from 'angular2/http';
+ * import { InMemoryBackendConfig, InMemoryBackendService, SEEDDATA } from '../in-memory-backend/in-memory-backend.service';
  * import { InMemoryStoryService } from '../api/in-memory-story.service';
  *
  * @Component({
  *   selector: ...,
  *   templateUrl: ...,
  *   providers: [
- *     HTTP_PROVIDERS,
+ *     HTTPPROVIDERS,
  *     provide(XHRBackend, { useClass: InMemoryBackendService }),
- *     provide(SEED_DATA, { useClass: InMemoryStoryService }),
+ *     provide(SEEDDATA, { useClass: InMemoryStoryService }),
  *     provide(InMemoryBackendConfig, { useValue: { delay: 600 } }),
  *   ]
  * })
@@ -117,24 +119,24 @@ export const isSuccess = (status: number): boolean => (status >= 200 && status <
 
 export class InMemoryBackendService {
 
-  protected _config: InMemoryBackendConfigArgs = new InMemoryBackendConfig();
-  protected _db: {};
+  protected config: InMemoryBackendConfigArgs = new InMemoryBackendConfig();
+  protected db: {};
 
   constructor(
-    @Inject(SEED_DATA) private _seedData: InMemoryDbService,
+    @Inject(SEED_DATA) private seedData: InMemoryDbService,
     @Inject(InMemoryBackendConfig) @Optional() config: InMemoryBackendConfigArgs) {
-    this._resetDb();
+    this.resetDb();
 
-    let loc = this._getLocation('./');
-    this._config.host = loc.host;
-    this._config.rootPath = loc.pathname;
-    Object.assign(this._config, config);
+    let loc = this.getLocation('./');
+    this.config.host = loc.host;
+    this.config.rootPath = loc.pathname;
+    Object.assign(this.config, config);
   }
 
-  createConnection(req: Request) {
-    let res = this._handleRequest(req);
+  createConnection(req: Request): Connection {
+    let res = this.handleRequest(req);
 
-    let response = new Observable((responseObserver: Observer<Response>) => {
+    let response = new Observable<Response>((responseObserver: Observer<Response>) => {
       if (isSuccess(res.status)) {
         responseObserver.next(res);
         responseObserver.complete();
@@ -144,8 +146,12 @@ export class InMemoryBackendService {
       return () => { }; // unsubscribe function
     });
 
-    response = response.delay(this._config.delay || 500);
-    return { response };
+    response = response.delay(this.config.delay || 500);
+    return {
+      readyState: ReadyState.Done,
+      request: req,
+      response
+    };
   }
 
   ////  protected /////
@@ -156,21 +162,25 @@ export class InMemoryBackendService {
    *
    * Expect URI pattern in the form :base/:collectionName/:id?
    * Examples:
-   *   api/characters
-   *   api/characters/42
-   *   api/characters.json/42   // ignores the ".json"
-   *   commands/resetDb  // resets the "database"
+   *   // for store with a 'characters' collection
+   *   GET api/characters          // all characters
+   *   GET api/characters/42       // the character with id=42
+   *   GET api/characters?name=^j  // 'j' is a regex; returns characters whose name contains 'j' or 'J'
+   *   GET api/characters.json/42  // ignores the ".json"
+   *
+   *   POST commands/resetDb  // resets the "database"
    */
-  protected _handleRequest(req: Request) {
-    let {base, collectionName, id, resourceUrl} = this._parseUrl(req.url);
-
+  protected handleRequest(req: Request) {
+    let {base, collectionName, id, resourceUrl, query} = this.parseUrl(req.url);
+    let collection = this.db[collectionName];
     let reqInfo: ReqInfo = {
       req: req,
       base: base,
-      collection: this._db[collectionName],
+      collection: collection,
       collectionName: collectionName,
       headers: new Headers({ 'Content-Type': 'application/json' }),
-      id: this._parseId(id),
+      id: this.parseId(collection, id),
+      query: query,
       resourceUrl: resourceUrl
     };
 
@@ -178,45 +188,73 @@ export class InMemoryBackendService {
 
     try {
       if ('commands' === reqInfo.base.toLowerCase()) {
-        options = this._commands(reqInfo);
+        options = this.commands(reqInfo);
 
       } else if (reqInfo.collection) {
         switch (req.method) {
           case RequestMethod.Get:
-            options = this._get(reqInfo);
+            options = this.get(reqInfo);
             break;
           case RequestMethod.Post:
-            options = this._post(reqInfo);
+            options = this.post(reqInfo);
             break;
           case RequestMethod.Put:
-            options = this._put(reqInfo);
+            options = this.put(reqInfo);
             break;
           case RequestMethod.Delete:
-            options = this._delete(reqInfo);
+            options = this.delete(reqInfo);
             break;
           default:
-            options = this._createErrorResponse(STATUS.METHOD_NOT_ALLOWED, 'Method not allowed');
+            options = this.createErrorResponse(STATUS.METHOD_NOT_ALLOWED, 'Method not allowed');
             break;
         }
 
       } else {
-        options = this._createErrorResponse(STATUS.NOT_FOUND, `Collection '${collectionName}' not found`);
+        options = this.createErrorResponse(STATUS.NOT_FOUND, `Collection '${collectionName}' not found`);
       }
 
     } catch (error) {
       let err = error.message || error;
-      options = this._createErrorResponse(STATUS.INTERNAL_SERVER_ERROR, `${err}`);
+      options = this.createErrorResponse(STATUS.INTERNAL_SERVER_ERROR, `${err}`);
     }
 
-    options = this._setStatusText(options);
-    if (this._config.defaultResponseOptions) {
-      options = this._config.defaultResponseOptions.merge(options);
+    options = this.setStatusText(options);
+    if (this.config.defaultResponseOptions) {
+      options = this.config.defaultResponseOptions.merge(options);
     }
 
     return new Response(options);
   }
 
-  protected _clone(data: any) {
+  /**
+   * Apply query/search parameters as a filter over the collection
+   * This impl only supports RegExp queries on string properties of the collection
+   * ANDs the conditions together
+   */
+  protected applyQuery(collection: any[], query: URLSearchParams) {
+    // extract filtering conditions - {propertyName, RegExps) - from query/search parameters
+    let conditions: {name: string, rx: RegExp}[] = [];
+    query.paramsMap.forEach((value: string[], name: string) => {
+      value.forEach(v => conditions.push({name, rx: new RegExp(decodeURI(v), 'i')}));
+    });
+
+    let len = conditions.length;
+    if (!len) { return collection; }
+
+    // AND the RegExp conditions
+    return collection.filter(row => {
+      let ok = true;
+      let i = len;
+      while (ok && i) {
+        i -= 1;
+        let cond = conditions[i];
+        ok = cond.rx.test(row[cond.name]);
+      }
+      return ok;
+    });
+  }
+
+  protected clone(data: any) {
     return JSON.parse(JSON.stringify(data));
   }
 
@@ -232,37 +270,37 @@ export class InMemoryBackendService {
    *   http.get('commands/config');
    *   http.post('commands/config', '{"delay":1000}');
    */
-  protected _commands(reqInfo: ReqInfo) {
+  protected commands(reqInfo: ReqInfo) {
     let command = reqInfo.collectionName.toLowerCase();
     let method = reqInfo.req.method;
     let options: ResponseOptions;
 
     switch (command) {
       case 'resetdb':
-        this._resetDb();
+        this.resetDb();
         options = new ResponseOptions({ status: STATUS.OK });
         break;
       case 'config':
         if (method === RequestMethod.Get) {
           options = new ResponseOptions({
-            body: this._clone(this._config),
+            body: this.clone(this.config),
             status: STATUS.OK
           });
         } else {
           // Be nice ... any other method is a config update
           let body = JSON.parse(<string>reqInfo.req.text() || '{}');
-          Object.assign(this._config, body);
+          Object.assign(this.config, body);
           options = new ResponseOptions({ status: STATUS.NO_CONTENT });
         }
         break;
       default:
-        options = this._createErrorResponse(
+        options = this.createErrorResponse(
           STATUS.INTERNAL_SERVER_ERROR, `Unknown command "${command}"`);
     }
     return options;
   }
 
-  protected _createErrorResponse(status: number, message: string) {
+  protected createErrorResponse(status: number, message: string) {
     return new ResponseOptions({
       body: { 'error': `${message}` },
       headers: new Headers({ 'Content-Type': 'application/json' }),
@@ -270,22 +308,22 @@ export class InMemoryBackendService {
     });
   }
 
-  protected _delete({id, collection, collectionName, headers /*, req */}: ReqInfo) {
+  protected delete({id, collection, collectionName, headers /*, req */}: ReqInfo) {
     if (!id) {
-      return this._createErrorResponse(STATUS.NOT_FOUND, `Missing "${collectionName}" id`);
+      return this.createErrorResponse(STATUS.NOT_FOUND, `Missing "${collectionName}" id`);
     }
-    let exists = this._removeById(collection, id);
+    let exists = this.removeById(collection, id);
     return new ResponseOptions({
       headers: headers,
-      status: (exists || !this._config.delete404) ? STATUS.NO_CONTENT : STATUS.NOT_FOUND
+      status: (exists || !this.config.delete404) ? STATUS.NO_CONTENT : STATUS.NOT_FOUND
     });
   }
 
-  protected _findById(collection: any[], id: number) {
+  protected findById(collection: any[], id: number | string) {
     return collection.find((item: any) => item.id === id);
   }
 
-  protected _genId(collection: any): any {
+  protected genId(collection: any): any {
     // assumes numeric ids
     let maxId = 0;
     collection.reduce((prev: any, item: any) => {
@@ -294,42 +332,54 @@ export class InMemoryBackendService {
     return maxId + 1;
   }
 
-  protected _get({id, collection, collectionName, headers}: ReqInfo) {
-    let data = (id) ? this._findById(collection, id) : collection;
+  protected get({id, query, collection, collectionName, headers}: ReqInfo) {
+    let data = collection;
+
+    if (id) {
+      data = this.findById(collection, id);
+    } else if (query) {
+      data = this.applyQuery(collection, query);
+    }
+
     if (!data) {
-      return this._createErrorResponse(STATUS.NOT_FOUND,
+      return this.createErrorResponse(STATUS.NOT_FOUND,
         `'${collectionName}' with id='${id}' not found`);
     }
     return new ResponseOptions({
-      body: { data: this._clone(data) },
+      body: { data: this.clone(data) },
       headers: headers,
       status: STATUS.OK
     });
   }
 
-  protected _getLocation(href: string) {
+  protected getLocation(href: string) {
     let l = document.createElement('a');
     l.href = href;
     return l;
   };
 
-  protected _indexOf(collection: any[], id: number) {
+  protected indexOf(collection: any[], id: number) {
     return collection.findIndex((item: any) => item.id === id);
   }
 
-  // tries to parse id as integer; returns input id if not an integer.
-  protected _parseId(id: string): any {
+  // tries to parse id as number if collection item.id is a number.
+  // returns the original param id otherwise.
+  protected parseId(collection: {id: any}[], id: string): any {
     if (!id) { return null; }
-    let idNum = parseInt(id, 10);
-    return isNaN(idNum) ? id : idNum;
+    let isNumberId =  collection[0] && typeof collection[0].id === 'number';
+    if (isNumberId) {
+      let idNum = parseFloat(id);
+      return isNaN(idNum) ? id : idNum;
+    }
+    return id;
   }
 
-  protected _parseUrl(url: string) {
+  protected parseUrl(url: string) {
     try {
-      let loc = this._getLocation(url);
-      let drop = this._config.rootPath.length;
+      let loc = this.getLocation(url);
+      let drop = this.config.rootPath.length;
       let urlRoot = '';
-      if (loc.host !== this._config.host) {
+      if (loc.host !== this.config.host) {
         // url for a server on a different host!
         // assume it's collection is actually here too.
         drop = 1; // the leading slash
@@ -339,22 +389,23 @@ export class InMemoryBackendService {
       let [base, collectionName, id] = path.split('/');
       let resourceUrl = urlRoot + base + '/' + collectionName + '/';
       [collectionName] = collectionName.split('.'); // ignore anything after the '.', e.g., '.json'
-      return { base, id, collectionName, resourceUrl };
+      let query = loc.search && new URLSearchParams(loc.search.substr(1));
+      return { base, id, collectionName, resourceUrl, query };
     } catch (err) {
       let msg = `unable to parse url '${url}'; original error: ${err.message}`;
       throw new Error(msg);
     }
   }
 
-  protected _post({collection, /* collectionName, */ headers, id, req, resourceUrl}: ReqInfo) {
+  protected post({collection, /* collectionName, */ headers, id, req, resourceUrl}: ReqInfo) {
     let item = JSON.parse(<string>req.text());
     if (!item.id) {
-      item.id = id || this._genId(collection);
+      item.id = id || this.genId(collection);
     }
     // ignore the request id, if any. Alternatively,
     // could reject request if id differs from item.id
     id = item.id;
-    let existingIx = this._indexOf(collection, id);
+    let existingIx = this.indexOf(collection, id);
     if (existingIx > -1) {
       collection[existingIx] = item;
       return new ResponseOptions({
@@ -366,22 +417,22 @@ export class InMemoryBackendService {
       headers.set('Location', resourceUrl + '/' + id);
       return new ResponseOptions({
         headers: headers,
-        body: { data: this._clone(item) },
+        body: { data: this.clone(item) },
         status: STATUS.CREATED
       });
     }
   }
 
-  protected _put({id, collection, collectionName, headers, req}: ReqInfo) {
+  protected put({id, collection, collectionName, headers, req}: ReqInfo) {
     let item = JSON.parse(<string>req.text());
     if (!id) {
-      return this._createErrorResponse(STATUS.NOT_FOUND, `Missing '${collectionName}' id`);
+      return this.createErrorResponse(STATUS.NOT_FOUND, `Missing '${collectionName}' id`);
     }
     if (id !== item.id) {
-      return this._createErrorResponse(STATUS.BAD_REQUEST,
+      return this.createErrorResponse(STATUS.BAD_REQUEST,
         `"${collectionName}" id does not match item.id`);
     }
-    let existingIx = this._indexOf(collection, id);
+    let existingIx = this.indexOf(collection, id);
     if (existingIx > -1) {
       collection[existingIx] = item;
       return new ResponseOptions({
@@ -391,15 +442,15 @@ export class InMemoryBackendService {
     } else {
       collection.push(item);
       return new ResponseOptions({
-        body: { data: this._clone(item) },
+        body: { data: this.clone(item) },
         headers: headers,
         status: STATUS.CREATED
       });
     }
   }
 
-  protected _removeById(collection: any[], id: number) {
-    let ix = this._indexOf(collection, id);
+  protected removeById(collection: any[], id: number) {
+    let ix = this.indexOf(collection, id);
     if (ix > -1) {
       collection.splice(ix, 1);
       return true;
@@ -410,11 +461,11 @@ export class InMemoryBackendService {
   /**
    * Reset the "database" to its original state
    */
-  protected _resetDb() {
-    this._db = this._seedData.createDb();
+  protected resetDb() {
+    this.db = this.seedData.createDb();
   }
 
-  protected _setStatusText(options: ResponseOptions) {
+  protected setStatusText(options: ResponseOptions) {
     try {
       let statusCode = STATUS_CODE_INFO[options.status];
       options['statusText'] = statusCode ? statusCode.text : 'Unknown Status';